moved remove_file_from_database into importer.py
